pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "loguru>=0.7.3",
    "pydantic>=2.10.6",
    "pydantic-ai-slim[openai]>=0.0.24",
    "python-dotenv>=1.0.1",
]
